backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.9
